requirements.txt
Django==1.11.4
django-taggit==0.22.1
flake8==3.4.1
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.2
